Chatbox now keeps focus after sending message. Also fixed steam avatars not being replicated.

+++ Source/ZombieGame/Public/LobbyGameMode.h
@@ -38,7 +38,7 @@ private:
 
 	void NotifyClientsOfPlayerLeave(int32 LeavingPlayerID);
 
-	int32 PlayerIDIterator;
+	uint8 PlayerIDIterator;
 
 	UPROPERTY()
 	TArray<class ALobbyPlayerController*> ConnectedPlayers;
